Add files via upload
The first 2 files which are named "breast-cancer-wisconsin "   are real dataset which are collected on 15th July 1992.   And the last 3 .py files are the codes which can make prediction over those data.  Since the data isn't very large we had used support vector machine classifier and it gave us fairly good accuracy over the data.  The codes are generally showing the acuuracy of each model over the data. Hope you like it.